custom_exception_try_catch

+++ class_1/generic_exceptions/src/main/java/org/example/Main.java
@@ -11,7 +11,7 @@ public class Main {
         }
     }
 
-    public static double add(double dividend, double divisor) {
+    public static int add(int dividend, int divisor) {
         return dividend / divisor;
     }
 }
